Bug fixes to selection
update to imagej 1.49v introduced errors into the selection feature of the program. ROIs could no longer be drawn and selections could not be made. These were fixed by moving when
we get the ROI information to before changing the image in the main window. Selections are now updated after releasing the mouse button.